Changed Fonts and Added UI Menu/Submenu
Changed all of the fonts to OpenSans font because it looks more pleasant than Ariel in my opinion (we can totally change it to whatever you all want). Additionally, I have added the basic template for how the UI menu/submenu will be. Later, we can properly label the buttons and link the buttons to the right scenes.